requirements.txt
asgiref==3.6.0
Django==4.0
djangorestframework==3.13.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.5.0
pytz==2022.7
sqlparse==0.4.3
pygments